requirements.txt
absl-py==0.11.0
apache-bookkeeper-client==4.10.0
appdirs==1.4.4
asn1crypto==0.24.0
astor==0.8.1
astunparse==1.6.3
attrs==20.3.0
audioread==2.1.9
backcall==0.2.0
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==2.1.4
cycler==0.10.0
decorator==4.4.2
fastavro==0.24.0
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.6
importlib-metadata==3.4.0
iniconfig==1.1.1
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
librosa==0.8.0
llvmlite==0.35.0
Markdown==3.3.3
matplotlib==3.3.4
numba==0.52.0
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.8
pandas==1.1.5
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
pooch==1.3.0
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.14.0
ptyprocess==0.7.0
pulsar-client==2.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
Pygments==2.8.0
pygobject==3.26.1
pymmh3==0.0.5
pyparsing==2.4.7
pytest==6.2.2
pytest-timeit==0.3.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
pyxdg==0.25
PyYAML==5.4.1
ratelimit==2.2.1
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7
scikit-learn==0.24.1
scipy==1.5.4
SecretStorage==2.3.1
six==1.15.0
SoundFile==0.10.3.post1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
tensorflow-io==0.17.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0